Argparser in cmaze and some fix

+++ src/app_config.cc
@@ -3,7 +3,7 @@
 
 void AppConfig::FillArgumentsList(argparse::ArgumentParser &appOptions)
 {
-   appOptions.add_argument("-s", "--seed").help("Random number generator seed").default_value(0).scan<'i', long>();
+   appOptions.add_argument("-s", "--seed").help("Random number generator seed").default_value((int)0).scan<'i', int>();
    appOptions.add_argument("-l", "--lines").help("Number of lines").default_value(66).scan<'i', int>();
    appOptions.add_argument("-c", "--columns").help("Number of columns").default_value(110).scan<'i', int>();
    appOptions.add_argument("-e", "--heuristic").help("Heuristic: 1-euclidean, 2-octagonal 3-manhattan").default_value(3).scan<'i', int>();
@@ -15,7 +15,7 @@ void AppConfig::ProcessArguments(argparse::ArgumentParser &appOptions)
    columns = appOptions.get<int>("--columns");
    lines = appOptions.get<int>("--lines");
    heuristic = appOptions.get<int>("--heuristic");
-   seed = appOptions.get<long>("--seed");
+   seed = appOptions.get<int>("--seed");
    debugMode = appOptions.get<bool>("--verbose") == true;
 }
